Move tests utils to improve IDE.
there were 2 files called utils.py - which confused IDEs (and therefore me).

Renameed tests/utils.py to tests/test_utils.py and changed all the imports.
